Modularization of Textures and made cc.Texture2D available at start
