+++ cocos2d/core/textures/index.js
@@ -1,0 +1,3 @@
+require('./CCTexture2D');
+require('./CCTextureAtlas');
+require('./CCTextureCache');
