+++ moduleConfig.json
@@ -73,8 +73,6 @@
             "cocos2d/core/base-nodes/CCAtlasNodeCanvasRenderCmd.js",
             "cocos2d/core/base-nodes/CCAtlasNodeWebGLRenderCmd.js",
 
-            "cocos2d/core/textures/TexturesWebGL.js",
-            "cocos2d/core/textures/TexturesPropertyDefine.js",
             "cocos2d/core/textures/CCTexture2D.js",
             "cocos2d/core/textures/CCTextureCache.js",
             "cocos2d/core/textures/CCTextureAtlas.js",
